# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * namFindz — Lost & Found / Missing Documents Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface AuthUser {
  id: string;
  email: string;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface RegisterUserBody {
  /**
   * @minLength 3
   * @maxLength 254
   */
  email: string;
  /**
   * @minLength 8
   * @maxLength 128
   */
  password: string;
  /**
   * @minLength 1
   * @maxLength 80
   */
  firstName?: string;
  /**
   * @minLength 1
   * @maxLength 80
   */
  lastName?: string;
}

export interface LoginUserBody {
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface AuthSessionResponse {
  user: AuthUser;
}

export const LogoutSessionResponseValue = {
  success: true,
} as const;
export type LogoutSessionResponse = typeof LogoutSessionResponseValue;

export interface RequestPasswordResetBody {
  /**
   * @minLength 3
   * @maxLength 254
   */
  email: string;
}

export const PasswordResetRequestResponseValue = {
  success: true,
} as const;
export type PasswordResetRequestResponse =
  typeof PasswordResetRequestResponseValue;

export interface ResetPasswordBody {
  /**
   * @minLength 16
   * @maxLength 256
   */
  token: string;
  /**
   * @minLength 8
   * @maxLength 128
   */
  password: string;
}

export const PasswordResetCompleteResponseValue = {
  success: true,
} as const;
export type PasswordResetCompleteResponse =
  typeof PasswordResetCompleteResponseValue;

export interface ErrorEnvelope {
  error: string;
}

export interface HealthStatus {
  status: string;
}

export type ReportKind = (typeof ReportKind)[keyof typeof ReportKind];

export const ReportKind = {
  lost: "lost",
  found: "found",
} as const;

export type ReportStatus = (typeof ReportStatus)[keyof typeof ReportStatus];

export const ReportStatus = {
  active: "active",
  matched: "matched",
  under_review: "under_review",
  claimed: "claimed",
  closed: "closed",
} as const;

export interface DocumentType {
  slug: string;
  label: string;
}

export interface Report {
  id: number;
  kind: ReportKind;
  status: ReportStatus;
  documentType: string;
  title: string;
  nameOnDocument?: string | null;
  description: string;
  location: string;
  eventDate: string;
  contactName: string;
  contactPhone?: string | null;
  contactEmail?: string | null;
  photoUrl?: string | null;
  reward?: string | null;
  createdAt: string;
}

export interface MatchSuggestion {
  report: Report;
  /** Similarity score 0..1 */
  score: number;
  reason: string;
}

export type ReportDetail = Report & {
  messageCount: number;
  suggestedMatches: MatchSuggestion[];
};

export interface CreateReportInput {
  kind: ReportKind;
  documentType: string;
  title: string;
  nameOnDocument?: string | null;
  description: string;
  location: string;
  eventDate: string;
  contactName: string;
  contactPhone?: string | null;
  contactEmail?: string | null;
  photoUrl?: string | null;
  reward?: string | null;
}

export interface Message {
  id: number;
  reportId: number;
  authorName: string;
  body: string;
  createdAt: string;
}

export interface StatsSummary {
  totalReports: number;
  activeLost: number;
  activeFound: number;
  successfulRecoveries: number;
  /** Percentage 0..100 */
  recoveryRate: number;
}

export interface DocumentTypeCount {
  documentType: string;
  label: string;
  lostCount: number;
  foundCount: number;
}

export type ActivityItemKind =
  (typeof ActivityItemKind)[keyof typeof ActivityItemKind];

export const ActivityItemKind = {
  match: "match",
  recovery: "recovery",
  new_report: "new_report",
} as const;

export interface ActivityItem {
  id: number;
  kind: ActivityItemKind;
  title: string;
  documentType: string;
  location: string;
  timestamp: string;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type ListReportsParams = {
  kind?: ReportKind;
  status?: ReportStatus;
  documentType?: string;
  /**
   * Free-text search across name, item details, and location
   */
  query?: string;
  location?: string;
  dateFrom?: string;
  dateTo?: string;
};

export type UpdateReportStatusBody = {
  status: ReportStatus;
};

export type FlagReportBody = {
  reason: string;
};

export type FlagReport200 = {
  success: boolean;
};

export type SendMessageBody = {
  authorName: string;
  body: string;
};