# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: namFindz — Lost & Found / Missing Documents Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /auth/register:
    post:
      tags: [Auth]
      operationId: registerUser
      summary: Create a new account with email + password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUserBody'
      responses:
        '201':
          description: Account created and session started.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionResponse'
        '400':
          description: Invalid registration details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Email already in use.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /auth/login:
    post:
      tags: [Auth]
      operationId: loginUser
      summary: Sign in with email + password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginUserBody'
      responses:
        '200':
          description: Session started.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionResponse'
        '400':
          description: Invalid login details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Invalid email or password.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /auth/logout:
    post:
      tags: [Auth]
      operationId: logoutSession
      summary: End the current session
      parameters:
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Session ended.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSessionResponse'
  /auth/request-password-reset:
    post:
      tags: [Auth]
      operationId: requestPasswordReset
      summary: Request a password reset link by email
      description: Always returns success to prevent account enumeration. If the email exists, a reset link is sent.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestPasswordResetBody'
      responses:
        '200':
          description: Reset request accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordResetRequestResponse'
        '400':
          description: Invalid email.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /auth/reset-password:
    post:
      tags: [Auth]
      operationId: resetPassword
      summary: Set a new password using a reset token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordBody'
      responses:
        '200':
          description: Password updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordResetCompleteResponse'
        '400':
          description: Invalid or expired token, or invalid password.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  parameters:
...
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
    AuthUser:
      type: object
      required: [id, email, firstName, lastName]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    RegisterUserBody:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          minLength: 3
          maxLength: 254
        password:
          type: string
          minLength: 8
          maxLength: 128
        firstName:
          type: string
          minLength: 1
          maxLength: 80
        lastName:
          type: string
          minLength: 1
          maxLength: 80
    LoginUserBody:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
    AuthSessionResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: '#/components/schemas/AuthUser'
    LogoutSessionResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    RequestPasswordResetBody:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          minLength: 3
          maxLength: 254
    PasswordResetRequestResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ResetPasswordBody:
      type: object
      required: [token, password]
      properties:
        token:
          type: string
          minLength: 16
          maxLength: 256
        password:
          type: string
          minLength: 8
          maxLength: 128
    PasswordResetCompleteResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...