# lib/api-zod/src/generated/types/passwordResetCompleteResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * namFindz — Lost & Found / Missing Documents Portal API
 * OpenAPI spec version: 0.1.0
 */

export const PasswordResetCompleteResponseValue = {
  success: true,
} as const;
export type PasswordResetCompleteResponse =
  typeof PasswordResetCompleteResponseValue;