# lib/api-zod/src/generated/types/passwordResetRequestResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * namFindz — Lost & Found / Missing Documents Portal API
 * OpenAPI spec version: 0.1.0
 */

export const PasswordResetRequestResponseValue = {
  success: true,
} as const;
export type PasswordResetRequestResponse =
  typeof PasswordResetRequestResponseValue;