# lib/api-zod/src/generated/types/requestPasswordResetBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * namFindz — Lost & Found / Missing Documents Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface RequestPasswordResetBody {
  /**
   * @minLength 3
   * @maxLength 254
   */
  email: string;
}