# lib/api-zod/src/generated/types/resetPasswordBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * namFindz — Lost & Found / Missing Documents Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface ResetPasswordBody {
  /**
   * @minLength 16
   * @maxLength 256
   */
  token: string;
  /**
   * @minLength 8
   * @maxLength 128
   */
  password: string;
}